% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_containers.R
\name{blob_containers}
\alias{blob_containers}
\title{Return blob containers}
\usage{
blob_containers(
  stage = c("dev", "prod"),
  sas = Sys.getenv("DSCI_AZ_SAS_DEV"),
  service = c("blob", "file")
)
}
\arguments{
\item{stage}{Store to access, either `prod` (default) or `dev`. `dev`}

\item{sas}{Shared access signature key to access the storage account or
blob. Default is set to use  dev stage sas key via an a env var named
"DSCI_AZ_SAS_DEV"}

\item{service}{Service to access, either `blob` (default) or `file.`}
}
\description{
Create endpoint URL to access Azure blob or file storage on either the
`dev` or `prod` stage from specified storage account
}
\examples{
# load project containers
containers <- blob_containers()
AzureStor::list_blobs(
  container = containers$projects,
  dir = "ds-contingency-pak-floods"
)

# You can also list as many containers as you want.
containers = blob_containers()

AzureStor::list_blobs(
  container = containers$global,
  dir = "raster/cogs"
)
}
