% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_utils.R
\name{blob_floodscan_catalogue}
\alias{blob_floodscan_catalogue}
\title{blob_floodscan_catalogue
return useful data.frame for filtering and pathing to floodscan cog urls
TODO: only set up for dev container so far.}
\usage{
blob_floodscan_catalogue(stage = c("dev", "prod"))
}
\arguments{
\item{stage}{Store to access, either `prod` (default) or `dev`. `dev`}
}
\value{
data.frame with columns `name`, `date`, and `vp` (virtual path),
  `isdir`, `blobtype`
}
\description{
blob_floodscan_catalogue
return useful data.frame for filtering and pathing to floodscan cog urls
TODO: only set up for dev container so far.
}
\examples{
library(cumulus)
catalogue <- blob_floodscan_catalogue()
}
