% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_utils.R
\name{blob_fs_check_dates}
\alias{blob_fs_check_dates}
\title{blob_floodscan_catalogue}
\usage{
blob_fs_check_dates(
  stage = c("dev", "prod"),
  start_date = "1998-01-01",
  end_date = Sys.Date() - 2
)
}
\arguments{
\item{stage}{Store to access, either `prod` (default) or `dev`. `dev`}

\item{start_date}{`character` start date to begin check (default = 1998-01-01)}

\item{end_date}{`character` end date to end check (default = Sys.Date()-2)}
}
\value{
if no missing date returns message "No dates missing" else returns
 data.frame catalogue with missing dates
}
\description{
blob_floodscan_catalogue
return useful data.frame for filtering and pathing to floodscan cog urls
TODO: only set up for dev container so far.
}
\examples{
library(cumulus)
blob_fs_check_dates()
}
