% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_read.R
\name{blob_load_admin_lookup}
\alias{blob_load_admin_lookup}
\title{blob_load_admin_lookup}
\usage{
blob_load_admin_lookup()
}
\value{
data.frame containing admin lookup data contained in parquet file
}
\description{
blob_load_admin_lookup
}
\examples{
library(cumulus)
blob_load_admin_lookup()
}
