% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fieldmaps_sf.R
\name{download_fieldmaps_sf}
\alias{download_fieldmaps_sf}
\title{Download and load adm0 boundaries}
\usage{
download_fieldmaps_sf(iso3, layer)
}
\arguments{
\item{iso3}{ISO3 code}
}
\description{
Downloads the adm0 shapefile from fieldmaps, because it sources OCHA boundaries
but standardizes column names. Then it loads the file in using `sf::st_read()`.
}
