% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fieldmaps_sf.R
\name{download_shapefile}
\alias{download_shapefile}
\title{Download shapefile and read}
\usage{
download_shapefile(url, layer = NULL)
}
\arguments{
\item{url}{URL to download}

\item{layer}{Layer to read}
}
\value{
sf object
}
\description{
Download shapefile to temp file, unzipping zip files if necessary. Deals with zipped
files like geojson or gpkg files as well as shapefiles, when the unzipping
returns a folder. The file is then read with `sf::st_read()`.
}
