% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_asi_tabular.R
\name{fao_asi_adm1_tabular}
\alias{fao_asi_adm1_tabular}
\title{get tabular ASI FAO data at admin 1 level by country}
\usage{
fao_asi_adm1_tabular(iso3)
}
\arguments{
\item{iso3}{`character` containing country iso3 code}
}
\value{
`data.frame` containing contents of FAO admin 1 csv from GIEWS
}
\description{
get tabular ASI FAO data at admin 1 level by country
}
\examples{
df_syria <- fao_asi_adm1_tabular("SYR")
}
