% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_vhi_tabular.R
\name{fao_vhi_adm1_tabular}
\alias{fao_vhi_adm1_tabular}
\title{get tabular Vegetative Health Index (VHI) data from
FAO GIEWS system at admin 1 level by country}
\usage{
fao_vhi_adm1_tabular(iso3)
}
\arguments{
\item{iso3}{`character` containing country iso3 code}
}
\value{
`data.frame` containing contents of FAO admin 1 csv from GIEWS
}
\description{
get tabular Vegetative Health Index (VHI) data from
FAO GIEWS system at admin 1 level by country
}
\examples{
df_somalia <- fao_vhi_adm1_tabular("SOM")
}
