% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_utils.R
\name{fs_date}
\alias{fs_date}
\title{fs_date floodscan date
utility function to parse dates from file names depending on where they are
stored and with what syntax.}
\usage{
fs_date(stage = c("dev", "prod"))
}
\arguments{
\item{stage}{Store to access, either `prod` (default) or `dev`. `dev`}
}
\value{
date parsing function
}
\description{
fs_date floodscan date
utility function to parse dates from file names depending on where they are
stored and with what syntax.
}
