% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_containers.R
\name{get_sas_key}
\alias{get_sas_key}
\title{get_sas_key
Convenience function get SAS key based on stage.
TODO: will eventually creat some credential registration step to make this
more flexible, but this will work well internally for now}
\usage{
get_sas_key(stage, key_syntax_v = 3, write_access)
}
\arguments{
\item{stage}{Store to access, either `prod` (default) or `dev`. `dev`}

\item{key_syntax_v}{`integer` indicating which version of key env var naming
to use. Default is set to 2 (this is a bit of hack for now)}
}
\value{
SAS key based on stage
}
\description{
get_sas_key
Convenience function get SAS key based on stage.
TODO: will eventually creat some credential registration step to make this
more flexible, but this will work well internally for now
}
