% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_contents.R
\name{list_contents}
\alias{list_contents}
\title{list container contents or contents within "sub-directory" of container}
\usage{
list_contents(
  container,
  dir,
  stage = "dev",
  sas = Sys.getenv("DSCI_AZ_SAS_DEV")
)
}
\arguments{
\item{container}{`character` container name}

\item{dir}{``character`     For list_blobs, a string naming the directory.
Note that blob storage does not support real directories; this argument
simply filters the result to return only blobs whose names start with
the given value.}

\item{stage}{`character` stage of the container (default = dev)}

\item{sas}{`character` SAS token}
}
\value{
`data.frame` of container/"sub-directory" contents
}
\description{
list container contents or contents within "sub-directory" of container
}
\examples{
library(cumulus)
list_contents("global", "raster/cogs")
}
