% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_data_sets.R
\name{load_cerf_from_blob}
\alias{load_cerf_from_blob}
\title{Load CERF allocation data from blob storage}
\usage{
load_cerf_from_blob(
  filename = "cerf/cerf_hdx_download.parquet",
  stage = "dev",
  container = "global",
  progress_show = TRUE
)
}
\arguments{
\item{filename}{Path to the file in the container
(default: `"cerf/cerf_hdx_download.parquet").}

\item{stage}{Store to access, either `"dev"` (default) or `"prod"`.}
}
\value{
Data frame of CERF allocation data.
}
\description{
Load CERF allocation data from blob storage
}
\examples{
\dontrun{
df_cerf <- load_cerf_from_blob()
}
}
