% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_data_sets.R
\name{load_emdat_from_blob}
\alias{load_emdat_from_blob}
\title{Load EM-DAT data from blob storage}
\usage{
load_emdat_from_blob(
  filename = "emdat/processed/emdat_all.parquet",
  stage = "dev",
  container = "global",
  progress_show = TRUE
)
}
\arguments{
\item{filename}{Path to the file in the container
(default: `"emdat/processed/emdat_all.parquet"`).}

\item{stage}{Store to access, either `"dev"` (default) or `"prod"`.}
}
\value{
Data frame of EM-DAT data.
}
\description{
Load EM-DAT data from blob storage
}
\examples{
\dontrun{
df_emdat <- load_emdat_from_blob()
}
}
