% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_oni.R
\name{load_oni}
\alias{load_oni}
\title{load_oni
Convenience function to load tidy ENSO ONI data}
\usage{
load_oni(show_col_types = TRUE)
}
\arguments{
\item{show_col_types}{`logical` from readr::read_table() whether or not to
print column types when the data.frame/table is loaded}
}
\value{
`data.frame`
}
\description{
load_oni
Convenience function to load tidy ENSO ONI data
}
\examples{
library(cumulus)
load_oni()
}
