% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{pg_con}
\alias{pg_con}
\title{pg_con - connect to postgres DB. Either `prod` or `dev` stage}
\usage{
pg_con(stage = "prod", key_syntax_v = 3, write = FALSE)
}
\arguments{
\item{stage}{db to access, either `prod` (default) or `dev`. `dev`}

\item{key_syntax_v}{`integer` 1 or 2 (default). 2 is the current env var
naming status}

\item{write_access}{`logical` indicating if write access is needed
default=FALSE}
}
\value{
`PqConnection` object
}
\description{
pg_con - connect to postgres DB. Either `prod` or `dev` stage
}
\examples{
\dontrun{
library(cumulus)
con <-  pg_con() # prod
DBI::dbListTables(con)
# connect to dev instead
con_dev <- pg_con(stage="dev")
DBI::dbListTables(con_dev)
}
}
