% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{pg_creds}
\alias{pg_creds}
\title{pg_creds}
\usage{
pg_creds(stage = "prod", key_syntax_v = 3, write_access = FALSE)
}
\arguments{
\item{stage}{db to access, either `prod` (default) or `dev`. `dev`}

\item{key_syntax_v}{`integer` 1, 2, or 3 (default). 3 is the latest env var
naming status. 1 & 2 will be deprecated.}

\item{write_access}{`logical` indicating if write access is needed
default=TRUE}
}
\value{
list containing named env vars to use for authentication
}
\description{
pg_creds
}
