% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas5.R
\name{pg_load_seas5_historical}
\alias{pg_load_seas5_historical}
\title{pg_load_sea5_historical}
\usage{
pg_load_seas5_historical(
  con = NULL,
  iso3,
  adm_level,
  adm_name,
  convert_units = TRUE
)
}
\arguments{
\item{con}{postgres connection class object. If none provided it will
connect to main by default}

\item{iso3}{`character` iso3 code for country}

\item{adm_level}{`numeric` administrative level}

\item{adm_name}{`character` admin name of interest}

\item{convert_units}{`logical` should we convert from m/s to mm/month
default = TRUE}
}
\value{
data.frame with historical seas5 forecast
}
\description{
pg_load_sea5_historical
}
\examples{
library(cumulus)
pg_load_seas5_historical(
  iso3 = "AFG",
  adm_level=1,
  adm_name = "Faryab"
)
}
