% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_az_file.R
\name{read_az_file}
\alias{read_az_file}
\title{Read a Parquet file stored on Microsoft Azure Data Storage blob or file
storage}
\usage{
read_az_file(
  file_path,
  sheet = NULL,
  container,
  sas_key = Sys.getenv("DSCI_AZ_SAS_DEV"),
  service = "blob",
  stage = "dev",
  endpoint_template = Sys.getenv("DSCI_AZ_ENDPOINT")
)
}
\arguments{
\item{file_path}{Name of the file to read, including directory prefixes
(`input/` or `output/`)and file extension, such as `.parquet`.}

\item{sheet}{`character` name of sheet (tab) if file is excel format (.xlsx).
If NULL (default) it reads
the first tab or ignores sheet argument}

\item{container}{`character` container name in the Azure storage.}

\item{sas_key}{`character` SAS key. Default set to env var naming convention
used for hdx-signals}

\item{service}{service in the Azure storage to read from, either "blob"
(default) or "file"}

\item{stage}{Stage in the Azure storage to read from, either "blob" (default)
or "file"}
}
\value{
Data frame.
}
\description{
Reads a file from the Azure Storage
The file is read based on its prefix. Currently, the only support is for
Apache Parquet, GeoJSON, CSV, and XLSX/XLS files, but other support will be
added as necessary
}
\details{
Function parsing is done based on file type:

* Apache Parquet: [arrow::write_parquet()].
* GeoJSON: [sf::st_read()]
* CSV: [readr::read_csv()]
* XLSX: [readxl::read_excel()]
}
\examples{
read_az_file(
  file_path = "nga-aa-cholera/public/processed/nga-adm2-populations-2023.csv",
  container = "projects"
)
}
