% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas5.R
\name{seas5_aggregate_forecast}
\alias{seas5_aggregate_forecast}
\title{seas5_aggregate_forecast}
\usage{
seas5_aggregate_forecast(
  df,
  value = "mean",
  valid_months = c(3, 4, 5),
  by = c("iso3", "pcode", "issued_date")
)
}
\arguments{
\item{df}{`data.frame` with seas5 forecast as formatted in pg db}

\item{value}{`character` name of column containing seas5 value to forecast
(default = 'mean')}

\item{valid_months}{`integer` vector w/ valid_months to aggregate (by sum)}

\item{by}{`character` vector to group calculations by}
}
\value{
data.frame containing aggregated forecast values.
}
\description{
seas5_aggregate_forecast
}
\examples{
library(cumulus)
df_seas5 <- pg_load_seas5_historical(
  iso3 = "AFG",
  adm_level=1,
  adm_name = "Faryab",
  convert_units = TRUE
)
df_seas5_mam <- seas5_aggregate_forecast(
  df_seas5,
  value = "mean",
  valid_months = c(3,4,5),
  by = c("iso3", "pcode","name","issued_date")
  )


}
