% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_az_file.R
\name{write_az_file}
\alias{write_az_file}
\title{Write data frame to Microsoft Azure Data Storage}
\usage{
write_az_file(
  x,
  name,
  container,
  service = "blob",
  stage = "dev",
  endpoint_template = Sys.getenv("DSCI_AZ_ENDPOINT"),
  sas_key = Sys.getenv("DSCI_AZ_SAS_DEV")
)
}
\arguments{
\item{x}{Data frame, simple features, or spatRaster}

\item{name}{Name of the file to write, including prefix (`input/` or `output/`)
and filetype `.parquet`.}

\item{blob}{Blob in the Azure storage to read from, either `prod`, `dev`, or `wfp`.}
}
\value{
Nothing, but file is written to the `hdx-signals` bucket.
}
\description{
A convenient file saver that saves the data frame to the specified
parquet file. Simply writes out the data frame based on the file extension and
uploads that to the MADS container using [AzureStor::upload_blob()].
Currently supports Apache Parquet, GeoJSON, COG, and CSV files.
}
\details{
Files written out based on file type:

* Apache Parquet: [arrow::write_parquet()]
* GeoJSON: [sf::st_write()]
* CSV: [readr::write_csv()]
* COG: [terra::write_raster()]
}
