% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_containers.R
\name{azure_endpoint_url}
\alias{azure_endpoint_url}
\title{Create the endpoint URL}
\usage{
azure_endpoint_url(stage = c("dev", "prod"), service = c("blob", "file"))
}
\arguments{
\item{stage}{Store to access, either `prod` (default) or `dev`. `dev`}

\item{service}{Service to access, either `blob` (default) or `file.`}
}
\description{
Create endpoint URL to access Azure blob or file storage on either the
`dev` or `prod` stage from specified storage accountd
}
