% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_containers.R
\name{load_containers}
\alias{load_containers}
\title{Return blob container}
\usage{
load_containers(
  containers = c("global", "projects"),
  sas = Sys.getenv("DSCI_AZ_SAS_DEV"),
  service = c("blob", "file"),
  stage = c("dev", "prod"),
  storage_account = "imb0chd0"
)
}
\arguments{
\item{containers}{`character` vector containing the name of container to
load (default, "global", "projects")}

\item{service}{Service to access, either `blob` (default) or `file.`}

\item{stage}{Store to access, either `prod` (default) or `dev`. `dev`}

\item{storage_account}{Storage account to access. Default is `imb0chd0`}

\item{sas_key}{Shared access signature key to access the storage account or
blob. Default is set to use  dev stage sas key via an a env var named
"DSCI_AZ_SAS_DEV"}
}
\description{
Create endpoint URL to access Azure blob or file storage on either the
`dev` or `prod` stage from specified storage account
}
\examples{
# load project containers
pc <- load_containers(containers = "projects")
AzureStor::list_blobs(
  container = pc$PROJECTS_CONT,
  dir = "ds-contingency-pak-floods"
)

# You can also list as many containers as you want.
pc <- load_containers(containers = c("global", "projects"))
AzureStor::list_blobs(
  container = pc$GLOBAL_CONT,
  dir = "raster/cogs"
)
}
