% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg.R
\name{pg_creds}
\alias{pg_creds}
\title{pg_creds}
\usage{
pg_creds(stage = "prod", key_syntax_v = 2, write_access = TRUE)
}
\arguments{
\item{stage}{db to access, either `prod` (default) or `dev`. `dev`}

\item{key_syntax_v}{`integer` 1 or 2 (default). 2 is the current env var
naming status}

\item{write_access}{`logical` indicating if write access is needed
default=TRUE}
}
\value{
list containing named env vars to use for authentication
}
\description{
pg_creds
}
